.windsurf/skills/task/.shared/scripts/include/pyproject.toml
[project]
name = "task-cli"
version = "0.1.0"
description = "CLI for task management in nunchuck"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
task = "task_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
